requirements.txt
fastapi
uvicorn
gspread
google-auth
python-multipart
jinja2
